refactor(client): moved request to its own internal package.
This complete the round of major refactoring effort around one of the
most complex function in this package, this difficult to maintain.

Complexity has now been significantly reduced, which allowed to spot and
fix a few minor bugs (leaked resources, limited mime matching rules).

The complete client request functionality is now parked under its own
internal package.

Tests: the complete refactoring cycle has been kept under the control of
a comprehensive test harness that explore many edge cases.
