+++ client/internal/request/consts_test.go
@@ -1,13 +1,14 @@
 // SPDX-FileCopyrightText: Copyright 2015-2025 go-swagger maintainers
 // SPDX-License-Identifier: Apache-2.0
 
-package client
+package request
 
 // Test-only constants shared across the package's *_test.go files,
 // extracted to satisfy the goconst linter.
 
 const (
-	operationID    = "getTasks"
-	taskOneContent = "task 1 content"
-	taskTwoContent = "task 2 content"
+	valBirdWatching = "Bird watching"
+	valJohn         = "John"
+	valOrganTrail   = "Organ trail"
+	valTom          = "Tom"
 )
